+++ src/wizard/actionSummary/SelectActionImpl.css
