+++ src/wizard/actionSummary/SelectActionImpl.graphql
