+++ src/wizard/provideActionParameters/InputParametersForm.tsx
@@ -35,20 +35,18 @@ function InputParametersForm({
   };
 
   return (
-    <>
-      <Form
-        showErrorList={false}
-        schema={schema}
-        formData={initData}
-        onSubmit={onSuccessSubmit}
-        liveValidate={true}
-      >
-        {/*TODO(https://github.com/capactio/backlog/issues/30): Remove after btn refactor*/}
-        <Button htmlType="submit" className="form-submit-btn" type="primary">
-          Next
-        </Button>
-      </Form>
-    </>
+    <Form
+      showErrorList={false}
+      schema={schema}
+      formData={initData}
+      onSubmit={onSuccessSubmit}
+      liveValidate={true}
+    >
+      {/*TODO(https://github.com/capactio/backlog/issues/30): Remove after btn refactor*/}
+      <Button htmlType="submit" className="form-submit-btn" type="primary">
+        Next
+      </Button>
+    </Form>
   );
 }
